Finally made it upload non corrupted files (using insane nonsense). I just have to publish so that I can revert to this if need be.
